+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Managesieve: Fix match-type selector (remove unsupported options) in delete header action (#9610)
+- Improve installer to fix confusion about disabling SMTP authentication (#9801)
 
 ## Release 1.6.10